split TextUi class formatting responsibility to Formatter
